+++ src/app/eth-memento/ethMementoPlugin.ts
@@ -7,8 +7,8 @@ import {blockBasicModule} from "app/shared/module/block/blockBasic/blockBasicMod
 import {blockTxsModule} from "app/shared/module/block/blockTxs/blockTxsModule";
 import {blockAdvancedModule} from "app/shared/module/block/blockAdvanced/blockAdvancedModule";
 import {blockLogsBloomModule} from "app/shared/module/block/blockLogsBloom/blockLogsBloomModule";
-import {EthPriceAdapter} from "app/shared/adapter/EthPriceAdapter";
 import {BlockListAdapter} from "app/shared/adapter/block/BlockListAdapter";
+import {EthPriceAdapter} from "app/shared/adapter/EthPriceAdapter";
 
 const ethMementoPlugin: IPlugin = {
     init(configData: unknown, api, logger, publicPath) {
@@ -26,7 +26,7 @@ const ethMementoPlugin: IPlugin = {
             ethSymbol
         }));
 
-        api.addDataAdapter("adapter://aleth.io/extended/block/details", new BlockDetailsAdapter(dataSource));
+        api.addDataAdapter("adapter://aleth.io/full/block/details", new BlockDetailsAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/prices/latest", new EthPriceAdapter(dataSource, logger));
         api.addModuleDef("module://aleth.io/block/basic", blockBasicModule);
         api.addModuleDef("module://aleth.io/block/txs", blockTxsModule(ethSymbol));